+ChipsFactory
